refactor: Restore utility directory structure
adapters/extractors/parsersディレクトリをutilityディレクトリに統合。
従来の構造を維持しながらfeat/agentの機能を取り込み。

- utility/user_resolver.rs: ユーザーID解決
- utility/extract_form_data.rs: フォームデータ抽出（get_user_select追加）
- utility/datetime_parser.rs: 日時パース（新規追加）
- utility/resource_parser.rs: リソースパース（新規追加）

すべてのimport文をutilityモジュールに修正。
