--- src/interface/slack/extractors/mod.rs
@@ -1,18 +1,0 @@
-//! データ抽出ユーティリティ
-//!
-//! Slackイベントからのデータ抽出とパース処理を提供します。
-//!
-//! ## 責務
-//!
-//! Slackから送信されるペイロード（View Submission、Block Actionsなど）には、
-//! ユーザーが入力したデータが複雑なネストされた構造で格納されています。
-//! このモジュールは、その複雑な構造から必要なデータを抽出する関数を提供します。
-//!
-//! Slack APIの`state`構造（`HashMap<BlockId, HashMap<ActionId, Value>>`）から
-//! 特定のaction_idに対応する値を取り出す処理を担当します。
-//!
-//! ## モジュール
-//!
-//! - `form_data`: フォームデータの抽出（テキスト入力、日付選択、セレクトメニューなど）
-
-pub mod form_data;
