+++ src/interface/slack/mod.rs
@@ -34,16 +34,14 @@
 //! | Messages API | `slack_client/messages.rs` |
 //! | Block Kit | `views/` |
 
-pub mod adapters;
 pub mod app;
 pub mod async_execution;
 pub mod block_actions;
 pub mod constants;
-pub mod extractors;
 pub mod gateway;
-pub mod parsers;
 pub mod slack_client;
 pub mod slash_commands;
+pub mod utility;
 pub mod view_submissions;
 pub mod views;
