+++ src/interface/slack/utility/user_resolver.rs
