+++ src/interface/slack/view_submissions/reservation.rs
@@ -3,12 +3,12 @@
 use crate::domain::aggregates::resource_usage::value_objects::TimePeriod;
 use crate::domain::aggregates::resource_usage::value_objects::resource::{Gpu, Resource};
 use crate::domain::ports::repositories::ResourceUsageRepository;
-use crate::interface::slack::adapters::user_resolver;
 use crate::interface::slack::app::SlackApp;
 use crate::interface::slack::constants::*;
-use crate::interface::slack::extractors::form_data;
-use crate::interface::slack::parsers::datetime::parse_datetime;
-use crate::interface::slack::parsers::resource::parse_device_id;
+use crate::interface::slack::utility::datetime_parser::parse_datetime;
+use crate::interface::slack::utility::extract_form_data as form_data;
+use crate::interface::slack::utility::resource_parser::parse_device_id;
+use crate::interface::slack::utility::user_resolver;
 use crate::interface::slack::views::modals::result;
 use slack_morphism::prelude::*;
 use tracing::{error, info};
