+++ src/interface/slack/view_submissions/update.rs
@@ -2,11 +2,11 @@
 
 use crate::domain::aggregates::resource_usage::value_objects::{TimePeriod, UsageId};
 use crate::domain::ports::repositories::ResourceUsageRepository;
-use crate::interface::slack::adapters::user_resolver;
 use crate::interface::slack::app::SlackApp;
 use crate::interface::slack::constants::*;
-use crate::interface::slack::extractors::form_data;
-use crate::interface::slack::parsers::datetime::parse_datetime;
+use crate::interface::slack::utility::datetime_parser::parse_datetime;
+use crate::interface::slack::utility::extract_form_data as form_data;
+use crate::interface::slack::utility::user_resolver;
 use crate::interface::slack::views::modals::result;
 use slack_morphism::prelude::*;
 use tracing::{error, info};
